[tool] triage report csv output
